form empresa

+++ database/migrations/2024_05_31_205855_create_contatos_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('contatos', function (Blueprint $table) {
             $table->id();
-            $table->integer('celular')->nullable();
-            $table->integer('whatsapp')->nullable();
+            $table->bigInteger('celular')->nullable();
+            $table->bigInteger('whatsapp')->nullable();
             $table->string('email')->nullable();
             $table->string('descricao')->nullable();
             $table->foreignId('usuario_id')->references('id')->on('users');
